Add RawBlock model
The idea (for now) is that a block importer service will walk the
blockchain and pull every block into the `raw_blocks` table.
Then a block extractor service will do the actual intelligence of
understanding the RawBlock's content and expanding it into the
appropriate table/s.

Doing it this way allows me to keep the importer very generic.
I don't have to know everything that I need to know up front.
In the future when I inevitably learn about some new (to me)
aspect of a block that I need to care about, I won't have to re-walk
the whole blockchain. I can just re-walk the `raw_blocks` table.

The trade off is that the database will (likely) be twice as big
because it'll be storing the raw version of a block and the
extracted block data in at least one table (but often multiple tables).

Maybe I'll regret doing it this way later. We'll see.