Add beginning of a seed script
This will connect to an Ethereum node and get the latest block. Not very exciting yet.

It uses a handful of default values… which uses Infura.io. If you use Infura, you’ll need to include
your Infura API key in your `.env` at the root of this Rails application. It will not get committed or
pushed, because it is ignored in `.gitignore`.

If you run this in a `production` environment, like on Heroku, you’ll need to set the actual
ENV variable. Your `.env` file will not be present or read in `production`.

If you have your own Ethereum node and want to use it or change any the default values,
just add those to your `.env` file and/or in your ENV variables in `production`.